Refactoring and trying to fix sidewinder.
I split the sidewinder into two functions, and added a bunch of
diagnostics to try to figure out what's going on.  The diagnostics
inside sidewinder I'll throw away, but I'll keep the ability to print
grids with coordinate labels.

I've figured out that the problem is that I'm iterating over cells
rather than coordinate pairs.  The cell sequences all pull from the
original grid, so while the algorithm is working, it sometimes works
with outdated cell maps that don't reflect links that have already been
made.

The next commit will be a partial merge of the coord-bias branch, which
includes different versions of `grid-rows` and `grid-cells` that return
sequences of coordinate pairs instead of cells.