Further decomposition of sidewinder.
The logic for deciding to end the current run of cells is inherently
a bit complex.  The way Jamis' Ruby version is coded seemed clear,
but when I coded it that way in Clojure it felt annoyingly complex
and like it was working against the rest of the code (in the sense
that the intermediate values I prepared for doing that test were the
opposite of what was needed for taking the next steps).  I wondered
if doing a De Morgan's Law inversion on it would feel better.

The logic was complex enough that I wrote a test to verify that the
inverted version was equivalent, and then rewrote `sidewinder-step`
to use the inverted version.  It *did* feel cleaner, and that helped
me to find ways to decompose the function into smaller pieces.

But then a funny thing happened: once the code was decomposed, the
test for ending the run was nicely isolated, and it no longer mattered
that it didn't mesh well with the subsequent steps.  And the original
version seems a little easier to understand in isolation, not to mention
that there's value in having it align well with the Ruby code in the
book.  So I switched it back.

This commit leaves both versions in (along with the test I used to
verify the correctess of the inverted version) just so they'll be in
the repo history.  The next commit will pull those out.