Added cell-neighbors function.
Looking through the Ruby code, I noticed that there was a Cell#neighbors
method that I had missed and neglected to implement as a function.  There
was no need for that function in my existing code, but an arity-2 version
where the caller supplies the directions of interest does make binary-tree
slightly nicer, so I added it.

Side note: since the cell-neighbors function invokes both `map` and
`filter` on the directions vector, I toyed with the following version
that uses transducers:

```clojure
(defn cell-neighbors [cell directions]
  (sequence (comp (map cell) (filter identity))
            directions))
```

But that seems slightly less readable to me, and with a vector that
will never be longer than four, there's no performance consideration.
So I went back to the old-style version.