Find and display solution route.
I'm not entirely happy with the way all of this is composed
in snergly.core.  Since the "domain objects" (grids and cells)
are maps, I'm happy with adding extra annotation attributes, like
:distances and :path.  But the way it's actually working seems
messy.  That might mean I'm doing it wrong, or it might mean that
I shouldn't be happy with that strategy at all.