Property-based tests for algorithms (start)
Before I go any farther with the animations and more analysis routines,
I want to make sure I'm not being deceived by bugs in the algorithms---
both the maze generation and the incremental result reporting and the
change tracking.  Sounds like a job for test.check.

I expected this to be more difficult than it was.  All of the most
important properties are done (although two algorithms fail one of
them in an innocuous way that just adds one redundant animation
frame).  It should be fairly easy to finish these up and then write
similar properties for the analysis algorithms.

These property checks are, however, slow.  We have to generate mazes
of various sizes, run the algorithms ... and in the case of the
properties that validate the incremental result reporting, we have
to capture each update and validate properties on each one (and
in some cases on the differences between them).