Convert maze-binary-tree to a lazy sequence generator.
Well, I've re-learned the first lesson of Clojure once again. It's all
about seqs, and be lazy if you can.

I started writing these algorithms using loop/recur, and that seemed OK,
because they needed to be iterative on potentially large mazees.  But
then when I wanted to animate them and needed to make them asynchronous,
core.async had to be layered in there, and that got messy.  Finally,
trying to debug the corner cases of find-distances so that it reported
an animation update only when the entire flood-front advanced by a step,
I realized what I was doing wrong.

If I reimplement these algorithms as lazy sequence generators, I can
separate the core.async stuff into a single utility function, and I can
handle those corner cases by layering in a transducer to remove the
redundant updates, for no extra runtime cost and significantly less
complexity.

This is step one, and note that this change makes it easy to collapse
maze-binary-tree from two functions into one.  It's a net complexity
win even compared to the original synchronous algorithm based on
loop/recur.