Finish converting everything to sequence generators.
(That's not quite true.  I've finished converting everything that
was alreadyworking under ClojureScript: the maze algorithms, plus
find-distances.  But still: now I'm back to the same point where I
was last Thursday, but with all of the complex parts greatly
simplified, all of the properties passing, and a lot more confidence
in the correctness of everything.  Now I can get the other analysis
algorithms working, or take aim at the optimized canvas drawing
algorithms, with confidence that I'm not being misdirected by bugs
elsewhere.)