Improve distance-shading color allocation.
Based on running some experiments, I've learned that
it's very rare for the maximum distance in a distances run
to be more than 80% of the number of cells in the grid.
And with my previous setting, I pretty frequently saw the
shading get to the darkest shade too soon.  This (80% of
the total number of cells as the expected maximum distance)
should be a much better setting.