Draw without annotated grid (and stroke path)
Animation of distance- and path-finding is much faster,
because we don't incur the expense of annotating the grid cells
each time (and we don't generate nearly as much garbage).

Additionally, the path is drawn as a stroke through the center
of cells (rather than by coloring the cells on the path) so that
it's easier to see.