Ensure that we render every frame.
Before this change, rendering would often skip frames, because
(apparently) steps of the algorithms were continuing to process
while the rendering was happening, and especially for large grids
the render would take long enough for several algorithm steps, and
the next frame would render all of the progress that had happened
since the previous frame.

With this change, we can force every frame to render.  That really
slows things down a lot, but it opens the way to optimized,
incremental rendering that will make the animation both fast and
smooth.