Optimized rendering! (Still rough.)
This code does proper optimized rendering of the mazes, only actually
drawing the cells that have changed.  In spite of the following bugs,
it's clear that the basics are right this time (and they've never
even been close on my previous attempts, because of the skipped
animation frames).

Bugs I see:

* The first maze generated after page load does the initial full-grid
  render properly; subsequent mazes don't.
* Looks like the first render of each distances pass is not
  happening; that cell stays white.
* With this strategy working now, I have to go back to guessing the
  expected maximum distance and doing all of the background shading
  based on that guess.
* I'm not calculating the size of the cell background correctly, so
  the mazes look goofy.