Only use the animation channel.
Turns out, once I'm using channels properly to communicate and
coordinate between the animation code and React, I don't actually
need the channel that spools the grid sequences; I can just use them
directly.

And that will be important for the back-and-forth buttons I'm planning,
because what actually needs to go in the application state for proper
backtracking is the entire sequence, not just the current grid.