Refactor and simplify animation using `chain-seqs`
There was a lot of fiddling and duplication in snergly.animation
due to the need to animate all of the intermediate steps and then
seed analysis phases with the final step of the previous sequence.
The `chain-seqs` function now does that in a fairly general way,
and nearly everything got simpler as a result.