Fix a bug: row/col fields were not being coerced to ints.
The rework of the app state and parser functions broke the way
updates to maze state were being coerced to integers if they represented
row or column dimensions.  I don't know what was wrong, because I realized
that there was an entirely different way to handle that, and I like this
way better.