Replace schema with clojure.spec
This is a big commit that removes Prismatic Schema and replaces it with clojure.spec throughout the code. The ClojureScript version is still in complete working order, and all of the tests pass just as they did before. The Clojure command-line version is not currently functional because I still need to propagate the effects of some earlier data structure refactorings into the Clojure-only portions of the code.

Although everything that was specified with schema is now specified to the same level of detail using clojure.spec, there is more I want to do, because clojure.spec does have a somewhat different philosophy, and it can do more:

* Rather than using the `:req-un` and `:opt-un` features to deal with my existing, un-namespaced keys, I went ahead and shifted to namespaced keys. But that makes destructuring and direct access of values in those maps at least a little clumsy in other namespaces (and especially in tests). I'm guessing that part of the idea of using namespaced keys is that (typically) only functions in that same namespace would have to use those keys. This means that I probably need to add more functions so that other namespaces won't have to mess with the namespaced keys.
* There are a few places where my specs could be strengthened (the mixed-mode distances map, plus destructured arguments in `fdef`s as well as `fdef`s for variable-arity functions).
* I'd like to see how many of the custom unit tests I could replace by automatically generated property-based tests.
* I'd like to see how many of the custom property-based tests I could replace by automatically generated ones.
* I'd like to simplify my hand-built generators if possible.