Add grid/cell-neighbor function
This is the first of perhaps several function additions to the grid
and image namespaces to better encapsulate the various maps that those
namespaces use, reducing the use of explicitly namespaced keywords by
other namespaces.

I'm not sure how important that actually is. The encapsulation seems
like a great idea, and the explicitly namespaced keywords seem clumsy.
But this won't help destructuring at all, so if I go as far as a strict
policy that the namespaced keywords are not allowed to be used outside
of the namespace (which is where I'd go if I wanted to be strict about
encapsulation) then that would prohibit any of the functions in other
namespaces from using destructuring on any of these maps.

On yet another hand, at the moment there's *very* little destructuring
of these maps outside of their primary namespaces. So maybe that will be
a characteristic of these kinds of things ... that the functions that need
to do destructuring on several different keys in the map are the ones that
live in the same namespace, and functions outside just won't need that kind
of access.