correct some terminological sloppiness
The term "coordinate" refers to the individual numbers that make up a point or position in a grid.  I've been sloppy and also used that term to refer to the *pair* of coordinates.  Time to get consistent.  A coordinate should not be "coord" everywhere. A dimension should be "dimen" everywhere. And a position should be "pos" everywhere, except when pluralized, when it's "positions".