store bounds (:max-pos) in cells, not neighbors
This is the first substantial change in the representation of cells and grids since the first few days of the project. We need to be able to find the neighbors of a cell easily, and to do that, we need to know whether a cell is on a border of the maze. That's easy for the north and west sides (if the corresponding cell coordinate is zero, the cell is at the boundary). But for the south and east sides, we need to also know the size of the grid. Ideally, I suppose, we would also pass the grid whenever we're looking for a cell's neighbors, but that turns out to be inconvenient in some cases.

Originally, cells would contain four fields containing their neighbors, and `make-cell` would simply populate those when the cell was created. That sufficed until I tried to use `clojure.spec` to specify the constraints on a valid cell strongly enough that they could be automatically checked. At that point, it seemed better to store the boundary row and column information in the cells instead, and dynamically calculate the neighbors when needed.

This means that the `::max-pos` field will be identical in every cell in a grid, which seems a bit yucky ... but each cell is actually smaller now, because instead of four fields each containing a cell position, now there's just one. And the specs are much simpler and more reasonable. *And* the slightly tricky logic of computing the neighbors is in its own function where it's easier to test, rather than being part of `make-cell`. The tradeoff for all this is that when processing a grid, there's a bit more computation.

(It's certainly an option, in the future, to remove `::max-pos` from the cells as well, and just pass in the grid along with the cell when we need to find the neighbors. But I think that would make the specs more complicated and/or less precise again.)